pyproject.toml
[project]
name = "provenance"
version = "0.1.0"
description = "Local-first personal CRM and memory tool"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2",
    "typer>=0.12",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "python-slugify>=8.0",
    "pymupdf>=1.27.1",
    "prompt-toolkit>=3.0.52",
    "mcp>=1.0",
]

[project.scripts]
provenance = "cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli", "ai", "core", "provenance"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "mkdocs>=1.6,<2.0",
    "mkdocs-material>=9.5",
]
